Fix inverted reload times
Still have to do a proper lookup of the actual reload animation times,
they are currently statically assumed at 2 seconds. Reload animations seem to
be based on "hold" type if that is a table and/or "anim_usage". This needs some
further research as there are a few inconsistencies in that system.